Move logic determining which topper to apply
The logic determining which topper to apply to an article has been moved
from `o-topper` and into it's own library, `n-map-content-to-topper`.

We've done this so we can use the `mapContentToTopper` helper on the
server-side in `next-article`. Also, it moves us closer to having
`o-topper` less tied to Customer Products content.