Remove image credit from failing demos
To unblock the release, we are temporarily removing the image credit
from the news package demo. This decision was made after a conversation
between myself, @GlynnPhillips and @notlee.

We're temporarily removing rather than fixing because in attempting to
fix it, the problem spiralled out of control. We think there is a bigger
discussion to be had about `o-toppers` and it's something Content
Innovation would like to look at after the new live blogs has launched.